# [UI Tests] - Removing unnecessary flows to speed up tests