### Description
I looked at the current `setup` implementation and found a few things that we could do to speed up the tests without changing the functionality. Since the `setup` step is part of every test case, the speed improvements from this would be for all test cases, existing and new. The changes are:
1. Remove the duplicated `app.dismissSavePasswordPrompt()` call on `proceedWithValidPassword()` (it's already called in `tryProceed()`)
2. Since `login()` returns `mySiteScreen()` the `.goToMySiteScreen()` step in some test cases was unnecessary and can be removed. Removed on 3 tests - `MainNavigationTests.swift`, `ReaderTests.swift` and `StatsTests.swift`
3. Update the `dismiss...` functions to not check for isLoaded first but to call the action immediately, this would save some unnecessary calls, see:

Before change - The check for elements on screen is repeated 3 times, first to check for the first element, then the whole screen, then the first element again before tapping the Skip button (Below logs for the first two checks, I somehow missed the last check...)
```
    t =    42.27s     Dismiss onboarding questions prompt if needed.
    t =    42.27s         Confirm first element from `expectedElementGetters` is loaded on screen UITestsFoundation.OnboardingQuestionsPromptScreen
    t =    43.31s             Checking `Expect predicate `isHittable == 1` for object "Skip" Button`
    t =    43.31s                 Find the "Skip" Button
    t =    43.39s         Confirm whole screen UITestsFoundation.OnboardingQuestionsPromptScreen is loaded
    t =    44.42s             Checking `Expect predicate `isHittable == 1` for object "Skip" Button`
    t =    44.42s                 Find the "Skip" Button
```

After change - Check if element is loaded and tap on it
```
    t =    25.96s     Dismiss onboarding questions prompt if needed.
2023-07-11 04:26:08921 [ℹ️][LoginEpilogueScreen.swift]:84 dismissOnboardingQuestionsPromptIfNeeded() -> Dismissing onboarding questions prompt...
    t =    25.96s         Confirm first element from `expectedElementGetters` is loaded on screen UITestsFoundation.OnboardingQuestionsPromptScreen
    t =    27.00s             Checking `Expect predicate `isHittable == 1` for object "Skip" Button`
    t =    27.00s                 Find the "Skip" Button
    t =    27.07s         Tap "Skip" Button
    t =    27.07s             Wait for com.automattic.jetpack to idle
    t =    27.10s             Find the "Skip" Button
```

### Testing
CI should be 🟢 and time taken to complete the test suite _should_ be faster than in `trunk`